feature: use normalized table
Change db structure to store values in a normalized table, where
repeating string values are stored as references to separate tables.
Original data structure is viewable via skdata view. The table is
a Timescale table. Keep normalized references in memory as Promises,
as the db operation to fetch the reference or write a new string
value is async. This makes the data transformation from deltas to
batch-writable format heavy on Promises and cumbersome, but gives
speed when used in batches.